избавился от лишней сущности, вынес базовый url из сетевого слоя
